Use either authlib-injector or system properties, not both
Resolves https://github.com/fn2006/PollyMC/issues/154

In 1.20+, authlib-injector wants to intercept
https://api.minecraftservices.com/publickeys [0], but if we override the
services server using -Dminecraft.api.services.host, the request does
not get intercepted.

That's not an issue for API servers that implement the /publickeys
route, but Ely.by does not. A decent solution is to not set these
-Dminecraft.api.*.host system properties when authlib-injector is
available, since authlib-injector expects authlib to make requests to
Mojang's API servers.

[0] https://github.com/yushijinhun/authlib-injector/commit/18a0ce266963e98c47b65df7a71450a857e100ef